auth-service/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-cors
Flask-JWT-Extended==4.3.1
requests==2.27.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
PyJWT==2.6.0
psycopg2-binary==2.9.1
pymongo==3.12.0
